# docs/api/openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Moimday API
  version: 1.3.0
  description: MVP REST API ??SSOT with api-contract-mvp.md (multi-group, 30 members)
servers:
  - url: http://localhost:3000/v1
    description: Local dev
  - url: https://api.Moimday.example/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GroupSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        isAdmin: { type: boolean }
        memberCount: { type: integer }
    MeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        phoneMasked: { type: string, nullable: true }
        autoReminderEnabled: { type: boolean }
        activeGroupId: { type: string, format: uuid, nullable: true }
        groupId: { type: string, format: uuid, nullable: true, description: activeGroupId alias }
        groups:
          type: array
          items: { $ref: '#/components/schemas/GroupSummary' }
        isGroupAdmin: { type: boolean }
...
paths:
...
  /users/me:
    get:
      tags: [users]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user with groups and active group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
    patch:
      tags: [users]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
                autoReminderEnabled: { type: boolean }
      responses:
        '200': { description: Updated }
    delete:
      tags: [users]
      security: [{ bearerAuth: [] }]
      responses:
        '202': { description: Deletion scheduled }
  /users/me/active-group:
    patch:
      tags: [users]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupId]
              properties:
                groupId: { type: string, format: uuid }
      responses:
        '200':
          description: Active group updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '403': { description: FORBIDDEN }
  /users/me/push-token:
    put:
      tags: [users]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [platform, token]
              properties:
                platform: { type: string, enum: [ios, android] }
                token: { type: string }
      responses:
        '200': { description: Registered }
...